Core: Created eCAL Get Config C++ API (#568)
