+++ ecal/core/src/io/ecal_memfile_pool.cpp
@@ -24,7 +24,7 @@
 #include <ecal/ecal.h>
 
 #include "ecal_def.h"
-#include "ecal_config_hlp.h"
+#include "ecal_config_reader_hlp.h"
 #include "pubsub/ecal_subgate.h"
 
 #include "ecal_memfile_pool.h"
