+++ ecal/core/src/readwrite/ecal_reader_iceoryx.cpp
@@ -22,7 +22,7 @@
 **/
 
 #include "ecal_def.h"
-#include "ecal_config_hlp.h"
+#include "ecal_config_reader_hlp.h"
 #include "ecal_global_accessors.h"
 #include "pubsub/ecal_subgate.h"
